Inline: Move top-level Transformation into separate source file
